Inline log rendering
